# sentinel/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Defentrax API
  description: |
    Control-plane REST API for Defentrax (security monitoring / lightweight SIEM).

    This specification documents **implemented** routes only (inventory from
    `sentinel/api/internal/server/server.go`). It does **not** invent endpoints.

    **Not exposed as REST (yet):**
    - No standalone `GET /api/v1/agents` — agents appear nested under `GET /api/v1/servers/{id}`.
    - No `/api/v1/metrics` — use `GET /api/v1/dashboard/stats`.

    **Page permissions:** panel routes require a `pages:*` key in addition to the
    data permission named on each operation. `pages:dashboard`, `pages:alerts`,
    `pages:alert_detail`, `pages:events`, `pages:servers`, `pages:server_detail`,
    `pages:rules`, `pages:notifications`, `pages:team`, `pages:roles`, `pages:audit`.
    `SUPER_ADMIN` and `ADMIN` bypass these checks. Plugins and API keys have no page key.

    Interactive UI: `GET /api/v1/docs` (Swagger UI). Raw spec: `GET /api/v1/openapi.yaml`.
  version: 0.2.0
  contact:
    name: Defentrax
servers:
  - url: http://localhost:8080
    description: Local default (API_PORT / PORT)
...
paths:
...
  /api/v1/rules:
    get:
      tags: [Rules]
      operationId: listRules
      summary: List detection rules
      description: "Requires `rules:read` and `pages:rules`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      responses:
        "200":
          description: Rules
          content:
            application/json:
              schema:
                type: object
                required: [rules]
                properties:
                  rules:
                    type: array
                    items:
                      $ref: "#/components/schemas/Rule"
    post:
      tags: [Rules]
      operationId: createRule
      summary: Create a custom detection rule
      description: "Requires `rules:write` and `pages:rules`. The rule is stored and loaded into the detection engine. Names and alert text must stay possibility-oriented. At least one match field is required."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description, severity]
              properties:
                name:
                  type: string
                description:
                  type: string
                severity:
                  type: string
                  enum: [info, low, medium, high, critical]
                source:
                  type: string
                category:
                  type: string
                event_type:
                  type: string
                message_contains:
                  type: string
                fields:
                  type: object
                  additionalProperties:
                    type: string
                threshold_count:
                  type: integer
                window_seconds:
                  type: integer
                group_by:
                  type: array
                  items:
                    type: string
                action_title:
                  type: string
                action_description:
                  type: string
      responses:
        "201":
          description: Created rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/rules/{id}:
    get:
      tags: [Rules]
      operationId: getRule
      summary: Get detection rule
      description: "Requires `rules:read` and `pages:rules`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      responses:
        "200":
          description: Rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Rules]
      operationId: patchRule
      summary: Enable or disable a rule
      description: "Requires `rules:write` and `pages:rules`. Send `enabled` to toggle a rule, or the same fields as create to edit a bundled or custom rule. The rule id is kept."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/notification-channels:
    get:
      tags: [Notifications]
      operationId: listNotificationChannels
      summary: List notification channels
      description: "Requires `notifications:read` and `pages:notifications`. Secrets are never returned."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      responses:
        "200":
          description: Channels
          content:
            application/json:
              schema:
                type: object
                required: [channels]
                properties:
                  channels:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationChannel"
    post:
      tags: [Notifications]
      operationId: createNotificationChannel
      summary: Create notification channel
      description: |
        Requires `notifications:write` and `pages:notifications`.
        Needs `SECRETS_ENCRYPTION_KEY` (32-byte) when storing secrets.
        `channel_type`: `discord` | `slack` | `email` | `webhook`.
      security:
        - SessionCookie: []
        - BearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationChannel"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannel"
        "400":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/v1/notification-channels/{id}:
    get:
      tags: [Notifications]
      operationId: getNotificationChannel
      summary: Get notification channel
      description: "Requires `notifications:read` and `pages:notifications`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      responses:
        "200":
          description: Channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannel"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Notifications]
      operationId: patchNotificationChannel
      summary: Update notification channel
      description: "Requires `notifications:write` and `pages:notifications`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchNotificationChannel"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannel"
        "404":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Notifications]
      operationId: deleteNotificationChannel
      summary: Delete notification channel
      description: "Requires `notifications:write` and `pages:notifications`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/Error"
...
  /api/v1/notification-rules:
    get:
      tags: [Notifications]
      operationId: listNotificationRules
      summary: List notification rules
      description: "Requires `notifications:read` and `pages:notifications`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      responses:
        "200":
          description: Rules
          content:
            application/json:
              schema:
                type: object
                required: [rules]
                properties:
                  rules:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationRule"
    post:
      tags: [Notifications]
      operationId: createNotificationRule
      summary: Create notification rule
      description: "Requires `notifications:write` and `pages:notifications`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationRule"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationRule"
  /api/v1/notification-rules/{id}:
    get:
      tags: [Notifications]
      operationId: getNotificationRule
      summary: Get notification rule
      description: "Requires `notifications:read` and `pages:notifications`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      responses:
        "200":
          description: Rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationRule"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Notifications]
      operationId: patchNotificationRule
      summary: Update notification rule
      description: "Requires `notifications:write` and `pages:notifications`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchNotificationRule"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationRule"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Notifications]
      operationId: deleteNotificationRule
      summary: Delete notification rule
      description: "Requires `notifications:write` and `pages:notifications`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sentinel_session
      description: "Server-side session cookie (name configurable via `SESSION_COOKIE_NAME`; default `sentinel_session`)."
    BearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: sent_
      description: "Personal API key (`Authorization: Bearer sent_…`)."
    BearerAgentToken:
      type: http
      scheme: bearer
      bearerFormat: sagt_
      description: "Agent token from enrollment (`Authorization: Bearer sagt_…`)."
  parameters:
    PathUUID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: unauthorized
        message:
          type: string
        request_id:
          type: string
...
    Rule:
      type: object
      required: [id, rule_id, name, enabled, severity, version, definition, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        rule_id:
          type: string
          description: Stable YAML rule id
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        severity:
          type: string
        version:
          type: integer
        definition:
          description: Rule definition JSON
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NotificationChannel:
      type: object
      required: [id, name, channel_type, config, has_secrets, enabled, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        channel_type:
          type: string
          enum: [discord, slack, email, webhook]
        config: {}
        has_secrets:
          type: boolean
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateNotificationChannel:
      type: object
      required: [name, channel_type]
      properties:
        name:
          type: string
        channel_type:
          type: string
          enum: [discord, slack, email, webhook]
        config: {}
        secrets:
          type: object
          properties:
            webhook_url:
              type: string
            smtp_password:
              type: string
            headers:
              type: object
              additionalProperties:
                type: string
        enabled:
          type: boolean
    PatchNotificationChannel:
      type: object
      properties:
        name:
          type: string
        config: {}
        secrets:
          type: object
          properties:
            webhook_url:
              type: string
            smtp_password:
              type: string
            headers:
              type: object
              additionalProperties:
                type: string
        enabled:
          type: boolean
    NotificationRule:
      type: object
      required: [id, name, enabled, min_severity, triggers, channel_ids, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        enabled:
          type: boolean
        min_severity:
          type: string
        triggers:
          type: array
          items:
            type: string
        channel_ids:
          type: array
          items:
            type: string
            format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateNotificationRule:
      type: object
      required: [name]
      properties:
        name:
          type: string
        enabled:
          type: boolean
        min_severity:
          type: string
        triggers:
          type: array
          items:
            type: string
        channel_ids:
          type: array
          items:
            type: string
            format: uuid
    PatchNotificationRule:
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
        min_severity:
          type: string
        triggers:
          type: array
          items:
            type: string
        channel_ids:
          type: array
          items:
            type: string
            format: uuid
...